+++ src/delete.ts
@@ -188,6 +188,7 @@ export async function deleteFilesIn(files: string[],
     const ALL_DIRS = await ME.getAllDirectories();
 
     const CANCELLATION_SOURCE = new vscode.CancellationTokenSource();
+    let targetSession: symbol | false = false;
     try {
         // cancel button
         let isCancelling = false;
@@ -243,9 +244,13 @@ export async function deleteFilesIn(files: string[],
             });
             
             cancelBtn.command = CANCEL_BTN_COMMAND_ID;
-            RESTORE_CANCEL_BTN_TEXT();
 
             cancelBtn.show();
+
+            targetSession = await deploy_targets.waitForOtherTargets(
+                target, cancelBtn,
+            );
+            RESTORE_CANCEL_BTN_TEXT();
         }
 
         const WAIT_WHILE_CANCELLING = async () => {
@@ -451,6 +456,10 @@ export async function deleteFilesIn(files: string[],
         DISPOSE_CANCEL_BTN();
 
         deploy_helpers.tryDispose(CANCELLATION_SOURCE);
+
+        deploy_targets.unmarkTargetAsInProgress(
+            target, targetSession
+        );
     }
 }